% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Integrating a Trapezoid or a Simpson}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{xValues}{A numeric vector of x-values}

\item{yValues}{A numeric vector of y-values}

\item{result}{A numeric object of the value of the area under curve}

\item{rule}{An instruction of whether to calculate the 'result' value using the Trapezoidal rule or Simpson's rule}
}
\value{
an object of class Trapezoid or Simpson containing
\item{name}{The candidate's name}
\item{delegatesWon}{Number of delegates the candidate has won}
\item{party}{The candidate's party}
\item{delegatesNeeded}{Number of delegates needed for the candidate's party nomination}
}
\description{
Creates an object of class Trapezoid or Simpson with user-provided inputs
}
\note{
This function creates a Candidate!
}
\examples{

cruzCandidate<-createCandidate("Ted Cruz", 135, "Republican")
}
\author{
Matt Malis \email{malis.matt@gmail.com}
}
\seealso{
\code{\link{totalNeeded}} \code{\link{propNeeded}}
}

