\name{integrateItPack-package}
\alias{integrateItPack-package}
\alias{integrateItPack}
\docType{package}
\title{
\packageTitle{integrateItPack}
}
\description{
\packageDescription{integrateItPack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{integrateItPack}
\packageIndices{integrateItPack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{integrateItPack}

Maintainer: \packageMaintainer{integrateItPack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
