% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateItPack-package.R
\docType{package}
\name{integrateItPack}
\alias{integrateItPack}
\alias{integrateItPack-package}
\title{integrateItPack}
\description{
The integrateItPack package takes a function and performs numeric integration using either the Trapezoidal rule or Simpson's rule.
}
\examples{


}
\author{
Jeff Ziegler: \email{jeffreyziegler@wustl.edu}
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}, \code{\link{integrateIt}}, \code{\link{tolTest}}
}

