% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/tolTest.R
\name{print}
\alias{print}
\alias{print,ANY-method}
\alias{tolTest}
\alias{print,ANY-method}
\title{Adding squared values}
\usage{
print(x, y, startandend, Rule, ...)

tolTest(fun, tolerance, rule, start, correct, ...)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}

\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class Squares containing
 \item{Trapezoid}{Simpson}{integreated value}

An object of class Squares containing
 \item{Trapezoid}{Simpson}{integreated value}
}
\description{
Finds the sum of squared numbers

Finds the sum of squared numbers
}
\note{
This is a very simple function

This is a very simple function
}
\examples{

myX <- c(20, 3) 
myY <- c(-2, 4.1) 
addSquares(myX, myY)

myX <- c(20, 3) 
myY <- c(-2, 4.1) 
addSquares(myX, myY)
}
\seealso{
\code{\link{print}}

\code{\link{print}}
}
\author{
Lim

Lim
}
