% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Calculate the numeric integral with a specified degree of tolerance}
\usage{
tolTest(inputFunc, a, b, tol = 1e-08, rule = "Trapezoid")
}
\arguments{
\item{inputFunc}{Function to input into \code{tolTest} (function).}

\item{a}{Number of delegates not yet allocated in a given candidate's party primary (numeric).}

\item{b}{Lower limit of integration (numeric).}

\item{tol}{The amount of tolerance between estimate and real integral (numeric).}

\item{rule}{Clarify which rule with be used, Simpson's Rule or the Trapezoidal Rule (character).}
}
\value{
Returns the initial inputs, the partition size (final n), and absolute error of estimate (list).
}
\description{
Function that increases the number of intervals until the estimated integral is within the specified tolerance of the correct answer.
}
\examples{

}
\author{
Jeff Ziegler
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}, \code{\link{integrateIt}}
}

