% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Squares-class}
\alias{initialize,Squares-method}
\alias{getSquares,Squares-method}
\title{A squared value object}
\description{
Object of class \code{} are created by the \code{addSquares} and \code{subtractSquares} functions
}
\details{
An object of the class `SquaresPack' has the following slots:
\itemize{
\item \code{Trapezoid} The added or subtracted squared values
\item \code{x} The first input
\item \code{y} the second input 
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
