% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Perform numeric integration}
\usage{
integrateIt(xVec, yVec, a, b, rule)
}
\arguments{
\item{xVec}{Vector containing x values (numeric vector).}

\item{yVec}{Vector containing x values (numeric vector).}

\item{rule}{Clarify which rule, Simpson's or Trapezoid, will be used (character).}
}
\value{
An object of class 'Candidate' containing
\itemize{
\item \code{xVec} Vector containing x values (numeric vector).
\item \code{yVec} Vector containing x values (numeric vector).
\item \code{a} Lower limit of integration (numeric).
\item \code{b} Upper limit of integration (numeric).
\item \code{rule} Clarify which rule, Simpson's or Trapezoid, will be used (character).
}
}
\description{
Function creates an object of class 'Simpson' or 'Trapezoid' and determines the numeric
integral performing using the specified rule.
}
\examples{

xTest <- as.numeric(c(1:10))
yTest <- rep(1, 10)
newTrap <- integrateIt(xVec = xTest, yVec = yTest, a = 1, b = 10, rule ="Trapezoid")
print(newTrap)

}
\author{
Jeff Ziegler
}
\seealso{
\code{\link{Simpson}} \code{\link{Trapezoid}}
}

