% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Integrating a Trapezoid or a Simpson}
\usage{
integrateIt(xValues, yValues, rule)
}
\arguments{
\item{xValues}{A numeric vector of x-values}

\item{yValues}{A numeric vector of y-values}

\item{rule}{An instruction of whether to calculate the 'result' value using the Trapezoidal rule or Simpson's rule; either 'Trap' or 'Simp' (default to 'Trap')}
}
\value{
an object of class Trapezoid or Simpson containing
\item{xValues}{A numeric vector of x-values}
\item{yValues}{A numeric vector of y-values}
\item{result}{A numeric object of the value of the area under the curve}
}
\description{
Creates an object of class Trapezoid or Simpson with user-provided inputs
}
\note{
This function calculates the area under a curve, two different ways!
}
\examples{

myIntegral<-integrateIt(myXvals, myYvals, 'Simp')
}
\author{
Matt Malis \email{malis.matt@gmail.com}
}
\seealso{
\code{\link{Trapezoid}} \code{\link{Simpson}}
}

