% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Simpson-method}
\alias{initialize,Simpson-method}
\alias{plot,}
\alias{print,Simpson-method}
\title{A Simpson object}
\description{
Objects of class \code{Simpson} are created through the \code{integrateIt} function. Objects of
this class include \code{plot} and \code{print} methods.
}
\details{
An object of the class 'Simpson' has the following slots:
\itemize{
\item \code{xVec} Vector containing x values (numeric vector).
\item \code{yVec} Vector containing x values (numeric vector).
\item \code{a} Lower limit of integration (numeric).
\item \code{b} Upper limit of integration (numeric).
\item \code{rule} Clarifying that Simpson's rule will be used (character).
}
}
\author{
Jeff Ziegler
}

