% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\title{A Simpson object}
\description{
X and Y values, and the area under the curve, calculated using the Simpson's rule
}
\details{
An object of the class 'Simpson' has the following slots:
\itemize{
\item \code{xValues} A numeric vector of x-values
\item \code{yValues} A numeric vector of y-values
\item \code{result} A numeric object of the value of the area under the curve
}
}
\author{
Matt Malis: \email{malis.matt@gmail.com}
}

