% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Trapezoid-method}
\alias{initialize,Trapezoid-method}
\alias{plot,}
\alias{print,Trapezoid-method}
\title{A Trapezoid object}
\description{
Objects of class \code{Trapezoid} are created through the \code{integrateIt} function. Objects of
this class include \code{plot} and \code{print} methods.
}
\details{
An object of the class 'Trapezoid' has the following slots:
\itemize{
\item \code{xVec} Vector containing x values (numeric vector).
\item \code{yVec} Vector containing x values (numeric vector).
\item \code{a} Lower limit of integration (numeric).
\item \code{b} Upper limit of integration (numeric).
\item \code{rule} Clarifying that the Trapezoid rule will be used (character).
}
}
\author{
Jeff Ziegler
}

