% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\title{A Trapezoid object}
\description{
X and Y values, and the area under the curve, calculated using the Trapezoidal rule
}
\details{
An object of the class 'Trapezoid' has the following slots:
\itemize{
\item \code{xValues} A numeric vector of x-values
\item \code{yValues} A numeric vector of y-values
\item \code{result} A numeric object of the value of the area under the curve
}
}
\author{
Matt Malis: \email{malis.matt@gmail.com}
}

